isolate business logic -> orderbook/resolvers
