+++ orderbook/api.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import resolver from "../resolver";
-import resolvers from "./resolvers";
+import * as resolvers from "./resolvers";
 /**
  * Allow to be mounted/partitioned '/api' |'/api/v1' ... etc
  */
