+++ orderbook/index.ts
@@ -1,4 +1,4 @@
 export { default as api } from "./api";
 export { default as reducer } from "./slice";
-export { default as resolvers } from "./resolvers";
+export * from "./resolvers";
 export * from "./types";
